backend/requirements.txt
flask

flask_cors
requests
python-dotenv
unstructured